refactor: completeWork while-loop 적용하여 리팩토링
Desc
- completeWork 내부의 재귀 호출 패턴을 제거하고 while-loop 패턴을 적용하여 리팩토링 진행
- beginWork 수행 중 fiber의 child가 더 이상 존재하지 않을 경우 completeWork를 수행하고 sibling 탐색 후, sibling 노드 역시 존재하지 않으면 parent 노드로 이동하도록 처리
